Remove core folder
