+++ lib/src/routing/ui/screens/books_tab.dart
@@ -1,23 +1,22 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
-
-import 'package:flutter_boolean_template/src/core/routing/router/app_router.dart';
+import 'package:flutter_boolean_template/src/routing/router/app_router.dart';
 
 @RoutePage()
-class SettingsScreen extends StatelessWidget {
-  const SettingsScreen({super.key});
+class BooksScreen extends StatelessWidget {
+  const BooksScreen({super.key});
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('Settings'),
+        title: const Text('Books'),
         leading: const AutoLeadingButton(),
       ),
-      backgroundColor: Colors.red,
+      backgroundColor: Colors.blue,
       body: Center(
         child: FilledButton(
           onPressed: () {
-            context.pushRoute(const SettingsDetailsRoute());
+            context.pushRoute(const BookDetailsRoute());
           },
           child: const Text('Push Details'),
         ),
