+++ lib/src/routing/ui/screens/settings_tab.dart
@@ -1,23 +1,22 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
-
-import 'package:flutter_boolean_template/src/core/routing/router/app_router.dart';
+import 'package:flutter_boolean_template/src/routing/router/app_router.dart';
 
 @RoutePage()
-class BooksScreen extends StatelessWidget {
-  const BooksScreen({super.key});
+class SettingsScreen extends StatelessWidget {
+  const SettingsScreen({super.key});
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('Books'),
+        title: const Text('Settings'),
         leading: const AutoLeadingButton(),
       ),
-      backgroundColor: Colors.blue,
+      backgroundColor: Colors.red,
       body: Center(
         child: FilledButton(
           onPressed: () {
-            context.pushRoute(const BookDetailsRoute());
+            context.pushRoute(const SettingsDetailsRoute());
           },
           child: const Text('Push Details'),
         ),
