+++ lib/src/store/store_example.dart
